moved NotificationCenter folder from ui/components to ui/components/layout/
