+++ ui/components/designs/lifecycle/DeployStepper.tsx
@@ -27,13 +27,13 @@ import { styled } from '@sistent/sistent';
 import { useTheme } from '@sistent/sistent';
 import { EnvironmentIcon } from '@sistent/sistent';
 import { useContext } from 'react';
-import { NotificationCenterContext } from '../../NotificationCenter';
+import { NotificationCenterContext } from '../layout/NotificationCenter';
 import { useEffect } from 'react';
 import { OPERATION_CENTER_EVENTS } from 'machines/operationsCenter';
 import { capitalize } from 'lodash';
 import FinishFlagIcon from '@/assets/icons/FinishFlagIcon';
 import { DeploymentSummaryFormatter } from './DeploymentSummary';
-import { SEVERITY } from '../NotificationCenter/constants';
+import { SEVERITY } from '../layout/NotificationCenter/constants';
 import EnvironmentModal from '../shared/Modal/EnvironmentModal';
 import { openViewScopedToDesignInOperator } from '@/utils/utils';
 import { useRouter } from 'next/router';
