+++ ui/components/designs/lifecycle/DeploymentSummary.tsx
@@ -1,12 +1,12 @@
 import { NOTIFICATIONCOLORS } from '@/themes/index';
 import { Box, Stack, Typography, styled, useTheme } from '@sistent/sistent';
 
-import { FormatStructuredData, TextWithLinks } from '../../DataFormatter';
-import { SEVERITY_STYLE } from '../../NotificationCenter/constants';
+import { FormatStructuredData, TextWithLinks } from '../DataFormatter';
+import { SEVERITY_STYLE } from '../layout/NotificationCenter/constants';
 import { ComponentIcon } from './common';
 import { Button } from '@sistent/sistent';
 import { ExternalLinkIcon, componentIcon } from '@sistent/sistent';
-import { ErrorMetadataFormatter } from '../../NotificationCenter/formatters/error';
+import { ErrorMetadataFormatter } from '../layout/NotificationCenter/formatters/error';
 
 import { openViewScopedToDesignInOperator, useIsOperatorEnabled } from '@/utils/utils';
 import { useRouter } from 'next/router';
