+++ ui/components/layout/NotificationCenter/constants.tsx
@@ -1,8 +1,8 @@
-import { NOTIFICATIONCOLORS } from '../../themes';
-import AlertIcon from '../../assets/icons/AlertIcon';
-import ErrorIcon from '../../assets/icons/ErrorIcon';
-import { Colors } from '../../themes/app';
-import ReadIcon from '../../assets/icons/ReadIcon';
+import { NOTIFICATIONCOLORS } from '../../../themes';
+import AlertIcon from '../../../assets/icons/AlertIcon';
+import ErrorIcon from '../../../assets/icons/ErrorIcon';
+import { Colors } from '../../../themes/app';
+import ReadIcon from '../../../assets/icons/ReadIcon';
 import { InfoIcon } from '@sistent/sistent';
 
 export const SEVERITY = {
