+++ ui/components/layout/NotificationCenter/formatters/RegistrantSummaryFormatter.tsx
