+++ ui/components/layout/NotificationCenter/formatters/academy_events.tsx
