+++ ui/components/layout/NotificationCenter/formatters/common.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { LaunchIcon, Typography } from '@sistent/sistent';
-import { TextWithLinks } from '../../DataFormatter';
+import { TextWithLinks } from '../../../DataFormatter';
 
 type TitleLinkProps = {
   href: string;
