+++ ui/components/layout/NotificationCenter/formatters/model_registration.tsx
