+++ ui/components/layout/NotificationCenter/formatters/pattern_dryrun.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FormatDryRunResponse } from '../../designs/lifecycle/DryRun';
+import { FormatDryRunResponse } from '../../../DesignLifeCycle/DryRun';
 import { formatDryRunResponse } from 'machines/validator/designValidator';
 import { ValidationResults } from '@/components/designs/lifecycle/ValidateDesign';
