+++ ui/machines/operationsCenter.ts
@@ -1,5 +1,5 @@
-import { SEVERITY_TO_NOTIFICATION_TYPE_MAPPING } from '@/components/NotificationCenter/constants';
-import subscribeEvents from '@/graphql/subscriptions/EventsSubscription';
+import { SEVERITY_TO_NOTIFICATION_TYPE_MAPPING } from '@/components/layout/NotificationCenter/constants';
+import subscribeEvents from '@/components/graphql/subscriptions/EventsSubscription';
 import { emit, fromCallback, setup, spawnChild, stopChild } from 'xstate';
 import { store } from '../store';
 import { pushEvent } from '@/store/slices/events';
